pyproject.toml
[project]
name = "strands"
version = "0.1.0"
description = "Strands Agents API - Comprehensive agent management with memory, budget analysis, and portfolio orchestration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core FastAPI dependencies
    "fastapi[standard]==0.115.6",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "python-dotenv==1.0.1",

    # Strands Agents SDK and tools with Gemini support
    "strands-agents[gemini]>=1.15.0",
    "strands-agents-tools>=0.2.14",

    # Memory backend dependencies (Lab 2)
    # FAISS for local vector storage (default)
    "faiss-cpu>=1.9.0",
    # mem0.ai for memory management
    "mem0ai>=0.1.0",
    # OpenSearch for AWS memory backend (optional)
    "opensearch-py>=2.7.0",

    # Data analysis (Lab 1 & 3)
    "pandas>=2.2.0",
    "numpy>=1.26.0",

    # Financial data (Lab 3)
    "yfinance>=0.2.0",

    # Additional dependencies
    "httpx>=0.28.0",
    "python-multipart>=0.0.20",

    # Logging and monitoring
    "python-json-logger>=3.2.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "ruff>=0.1.0",           # Fast Python linter and formatter
]


[tool.ruff]
# General settings
line-length = 88
indent-width = 4
target-version = "py313"

# Files to exclude from linting
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".env",
    ".env.example",
]

[tool.ruff.lint]
# Select rule sets
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "W",   # pycodestyle warnings
    "C90", # mccabe complexity
    "I",   # isort import sorting
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "T10", # flake8-debugger
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "FURB", # refurb
    "RUF", # ruff-specific rules
    "PLC", # pylint conventions
    "PLE", # pylint errors
    "PLW", # pylint warnings
]

# Ignore specific rules
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Do not perform function calls in argument defaults
    "C901",  # Too complex (use discretion)
    "PLR0913", # Too many arguments
    "PLR0915", # Too many statements
    "PLW0603", # Using the global statement
    "T201",   # print found (allow for debugging)
]

# Allow automatic fixes
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Per-file ignores
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]  # Allow unused imports and imports not at top
"app/main.py" = ["PLW0603"]  # Allow global statement in main module
"docs/**/*" = ["ALL"]  # Ignore documentation files

[tool.ruff.format]
# Formatting settings
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Format docstring code examples
docstring-code-format = true
docstring-code-line-length = "dynamic"

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]

[tool.ruff.lint.mccabe]
# Maximum complexity for functions
max-complexity = 10

[tool.ruff.lint.isort]
# Import sorting settings
known-first-party = ["app"]
known-third-party = ["fastapi", "pydantic", "boto3", "pandas", "numpy", "yfinance", "mem0", "strands"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
force-single-line = false
force-sort-within-sections = true

[tool.ruff.lint.pyupgrade]
# pyupgrade settings
keep-runtime-typing = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
multiline-quotes = "double"

[tool.ruff.lint.pylint]
# Pylint settings
max-args = 7
max-locals = 15
max-returns = 6
max-branches = 12
max-statements = 50
max-public-methods = 20
